cdv: Find end of 'Last calls' section
This is a bug fix: the end of the 'Last calls' section is meant to be
an empty line, but in some situations this is not the case. This could
cause the complete dump to be read and stored as 'Last calls', causing
very unexpected behavior in the crashdump viewer.

This commit adds a check for ": ", which indicates that a new heading
has occured and thus the 'Last calls' section is completly read.