Check the right side of a transformation better
The right side of a transformation must be either a single call
to a transformation function OR a list of instructions. Mixing
them like this is not supported:

  some_instruction A B C => gen_something(A) | other B C

Unfortunately, beam_makeops would silenty ignore anything after the
function call, basically handling it in the same way as:

  some_instruction A B C => gen_something(A)

Add a sanity check to reject such mixed right-hand sides.