Implement escaping of special characters in wildcards
Allow characters with special meaning to be escaped using
\ (which must be writen as \\ in a string). That allows
matching of filenames containing characters that are special
in wildcards.

This is an incompatible change, but note that the use of backslashes
in wildcards would already work differently on Windows and Unix. Take
for example this call:

     filelib:wildcard("a\\b")

On Windows, filelib:wildcard/1 would look for a directory named
"a", and a file or directory named "b" inside it.

On Unix, filelib:wildcard/1 would look for a file named "a\\b".

With this commit applied, filelib:wildcard/1 will look for
a file named "ab" on both Windows and Unix.

https://bugs.erlang.org/browse/ERL-451