kernel: Don't call other modules from simple error logger
error_logger implements a simple handler which is used when no other
handler is active. This handler is used during system startup, when
there is no guarantee that all modules are loaded and must therefore
not use any other modules for the formatting.

Commit 41b856a2cb270ef807beaa84afe1bb890de5f237 introduced a call to
io_lib:printable_list/1, which is now removed.