Refactor macros for accessing Beam instructions
The BeamOp() macro in erl_vm.h is clumsy to use. All users
cast the return value to BeamInstr.

Define new macros that are easier to use. In the future,
we might want to pack an operand into the same word as
the pointer to the instruction, so we will define two macros.

BeamIsOpCode() is used to rewrite code like this:

  if (Instr == (BeamInstr) BeamOp(op_i_func_info_IaaI) {
    ...
  }

to:

  if (BeamIsOpCode(Instr, op_i_func_info_IaaI)) {
    ...
  }

BeamOpCodeAddr(op_apply_bif) is used when we need the address
for an instruction.

Also elimiminate the global variables em_* in beam_emu.c.
They are not really needed. Use the BeamOpCodeAddr() macro
instead.