Eliminate unnecessary and inconsistent casts
Consider the types in the code below:

    BeamInstr* I;
      .
      .
      .
    BeamInstr* next;
    next = (BeamInstr *) *I;
    Goto(next);

This is illogical. If 'I' points to a BeamInstr, then 'next' should
be a BeamInstr, not a pointer to a BeamInstr. The Goto() macros does
not require a pointer, because it will cast its argument to a void*
anyway.

Therefore, this code example can be simplified to:

    BeamInstr* I;
      .
      .
      .
    BeamInstr next;
    next = *I;
    Goto(next);

Similarly, we can remove the casts in the macros when NO_JUMP_TABLE
is defined.