Optimize packing for "optional use" operands
Operands that are marked with a "?" are not always used when
an instruction is executed. Enhance the packing algorithm to
place optional use operands into the instruction word even they
are not the first operand (as long as the total instruction
size stays the same).

Here are the instructions that will be packed differently
because of this change:

    allocate_heap t I t?
    allocate_heap_zero t I t?
    test_heap I t?

    i_bs_get_integer_8 x f? x
    i_bs_get_integer_16 x f? x
    i_bs_get_integer_32 x f? x