Fix zlib merge snafu
I accidentally merged a few fixes directly to maint-20 (and then to maint
and master), instead of merging to maint and master and waiting for a
patch to merge it into maint-20, leaving the related tickets dangling.

This commit does nothing beyond getting the tickets (and their release
notes) into the next patch.