Slightly speed up try/catch
The try_end and try_case instructions are implemented the same way
(try_case is translated to try_end by the loader).

We can do better than that. We know that try_case will only be executed
when an exception has been caught. Therefore, we know that x(0) is
the non-value and that x(1) through x(3) need to be shifted down to
x(0) through x(2). There is no need to test x(0) before shifting down.

try_end does not need the register shifting code at all.