Fix incorrect assumption about process activity
The tests assume that the most active process will be the current
one, which is no longer true since the delayed_write option now
uses a wrapper process for much of its work.

The timeout for this test has been increased to account for the
lack of delayed_write; 60s was enough for everything except the
debug build on some machines.