jinterface: Skip tests when hostname cannot be resolved
Check that the hostname can be resolved by the native resolver.
What normally has happened when it cannot is that gethostname()
returned a fqdn and `hostname -s` is not part of /etc/hosts.
This is solved on the erlang side by adding `hostname -s` to inet_db,
but java does not have a similar mechanism, so it fails when
it tries to connect to `hostname -s`.

This caused jinterface tests to fail when run in such an environment,
and travis-ci recently started doing this.