erl_docgen: Insert an anchor for "anchor" attribute
Certain functions, particularly in erlang(3), have specs with many
clauses, some of which are documented separately. No anchors were
inserted automatically for clauses (as is the case for functions), why
manually written anchors had to be added to the XML files. However,
these anchors had to be placed inside the description, and the links
would point into the text, not into the name of the function/clause.

The new "anchor" attribute is to be used where manually written
anchors have hitherto been used. The "anchor" attribute can also be
used without the "clause_i" attribute.