Correct erlang:is_builtin/3 for apply/2 and yield/0
erlang:is_builtin(erlang, M, F) returns false for apply/2 and
yield/0. The documentation for erlang:is_builtin/3 says that it
returns true for BIFs that are implemented in C. apply/2 and
yield/0 are implemented in C (as BEAM instructions), and
therefore the correct return value is true.

Also see a similar argument that was made for apply/3 in the past:

    http://erlang.org/pipermail/erlang-bugs/2015-October/005101.html

https://bugs.erlang.org/browse/ERL-500