Update file performance advice
The parts relating to drivers/ports are now obsolete, and the
provided example was far noisier than it had to be; the only
relevant metric is the number of calls and it's up to the user to
decide how those will be reduced. One could argue for its complete
removal, but I'm inclined to leave it be.