[ct] Remove call to application:which_applications()
The cth_log_redirect hook calls this function to check if the sasl
application is started. This is done for each error logger event.

In most systems, this is not a big problem, but on native compiled
code this call can be very slow if the message queue is long. This is
because huge message queue optimization is not implemented for native.