erl_parse: Use a new pat_expr rule for pattern expressions
In the grammar file for the Erlang language, patterns are parsed
as expressions. erl_lint will then weed out expressions that are
not legal patterns.

The rule sharing causes problems if we were to introduce new
syntax, for example a pattern followed by a ':'. There would
be a shift/reduce conflict, and a pattern followed by a ':'
would be parsed as a remote call.

Introduce a new pat_expr rule to express exactly the subset
of expressions that is allowed in pattern.

Note: For the moment, we must allow full expressions in
case clauses to keep 'merl' working.