Fix broken to_dis after merge from maint to master
d8d07a7593d811 that added the to_dis option to the compiler no longer
works when merged to master.  That is because of 79f28cfd8df1b7
that removed some unused code in erts_debug.

Fix this by adding a new function erts_debug:dis_to_file/2 and
use it from compile module where we have access to the filename
(in beam_listing we only have an opened file).