Fix purging of modules with "fake literals"
When compiling Erlang source code, the literal area for the
module can only contain data types that have a literal
syntax.

However, it is possible to sneak in other data types
(such as references) in the literal pool by compiling from
abstract or assembly code. Those "fake literals" would work
fine, but would crash the runtime system when the module containing
the literals was purged.

Although fake literals are not officially supported, the
runtime should not crash when attempting to use them.
Therefore, fix the garbage collection of literals and releasing
of literal areas.

https://bugs.erlang.org/browse/ERL-508