Don't build a stacktrace if it's only passed to erlang:raise/3
Consider the following function:

    function({function,Name,Arity,CLabel,Is0}, Lc0) ->
      try
        %% Optimize the code for the function.
      catch
        Class:Error:Stack ->
          io:format("Function: ~w/~w\n", [Name,Arity]),
          erlang:raise(Class, Error, Stack)
      end.

The stacktrace is retrieved, but it is only used in the call
to erlang:raise/3. There is no need to build a stacktrace
in this function. We can avoid the building if we introduce
an instruction called raw_raise/3 that works exactly like
the erlang:raise/3 BIF except that its third argument must
be a raw stacktrace.