v3_codegen: Remove check of operand for bs_context_to_binary
The bs_context_to_binary instruction only allows a register operand.
v3_codegen has a test to ensure that the operand is a register.

That test is no longer necessary. There used to be a possibility
that optimizations in sys_core_fold and the inliner could change
the operand for bs_context_to_binary to a binary literal. Since
09112806c15a81b that can no longer happen, because no more
optimizations are run after the introduction of the
bs_context_to_binary instruction.