Stop trying to maximize the use of x(0)
X register 0 used to be mapped to a hardware register, and therefore
faster than the other registers. Because of that, the compiler
tried to use x(0) as much as possible as a temporary register.

That was changed a few releases ago. X register 0 is now placed
in the array of all X registers and has no special speed
advantage compared to the other registers.

Remove the code in the compiler that attempts to use x(0) as
much as possible. As a result, the following type of instruction
will be much less frequent:

 {put_list,Src,{x,0},{x,0}}

Instead, the following type of instruction will be more frequent:

 {put_list,Src,{x,X},{x,X}}

(Where X is an arbitrary X register.)

Update the runtime system to specialize that kind of put_list
instruction.