v3_codegen: Avoid excessive stack frame allocation
A 'case' or 'if' that does not occur last in a function clause will
always force a stack frame. The reasoning behind this is that in most
uses of 'case' there will be a function call from within the
'case'. When there is a function call, the stack frame is needed both
to save the continuation pointer and to save any X registers that will
need to survive the call.

When there is no function call from a 'case', the resulting stack
frame is annoying. There will be register shuffling, and the existence
of the stack frame may thwart many optimizations (for example, in
beam_dead).

Therefore, add an extra pass to v3_codegen to avoid creating a
stack frame when not needed.

https://bugs.erlang.org/browse/ERL-514