jinterface: add package name for Java 9
The Java 9 module system requires a module name to be defined. If none is provided, a default one is used derived from the JAR name, but that one is not very unique and can create problems. The recommended module name is the longest common package name for the source code included.