v3_codegen: Don't let exit BIFs force a stack frame
This is an enhancement of the optimization added in 2e5d6201bb044,
where we tried to avoid forcing a stack frame for functions
that don't really need them.

That optimization would not suppress the stack frame for this
function:

f(A) ->
    Res = case A of
              a -> x;
              b -> y
          end,
    {ok,Res}.

The reason is that internally the compiler would rewrite
the code to something like this:

f(A) ->
    Res = case A of
              a -> x;
              b -> y;
	      Other -> error({case_clause,Other})
          end,
    {ok,Res}.

The call to error/1 would force creation of a stack frame,
even though it is not really needed because error/1 causes
an exception.

Handle calls to exit BIFs specially to allow suppressing the
stack frame.